applib/light: add light_set_color_rgb888() for full tint
