+++ src/fw/syscall/syscall.h
@@ -195,7 +195,7 @@ void sys_light_enable_interaction(void);
 void sys_light_enable(bool enable);
 void sys_light_enable_respect_settings(bool enable);
 void sys_light_reset_to_timed_mode(void);
-void sys_light_set_color(uint8_t argb);
+void sys_light_set_color_rgb888(uint32_t rgb);
 void sys_light_set_system_color(void);
 
 bool sys_mobile_app_is_connected_debounced(void);
