+++ tools/generate_native_sdk/exported_symbols.json
@@ -4,7 +4,7 @@
               "You should also make sure you are obeying our API design guidelines:",
               "https://pebbletechnology.atlassian.net/wiki/display/DEV/SDK+API+Design+Guidelines"
             ],
-  "revision" : "95",
+  "revision" : "96",
   "version" : "2.0",
   "files": [
     "fw/drivers/ambient_light.h",
@@ -4522,6 +4522,11 @@
               "name": "light_set_color",
               "implName": "app_light_set_color",
               "addedRevision": "94"
+            }, {
+              "type": "function",
+              "name": "light_set_color_rgb888",
+              "implName": "app_light_set_color_rgb888",
+              "addedRevision": "96"
             }, {
               "type": "function",
               "name": "light_set_system_color",
